Split raw.py into several files
First suggestion of refactoring
