+++ src/mrpro/data/__init__.py
@@ -1,0 +1,45 @@
+from mrpro.data._AcqInfo import AcqInfo
+from mrpro.data._KData import KData
+from mrpro.data._KHeader import KHeader
+
+# Valid acquisition flags for readouts.
+ACQ_FLAGS = (
+    'ACQ_NO_FLAG',
+    'ACQ_FIRST_IN_ENCODE_STEP1',
+    'ACQ_LAST_IN_ENCODE_STEP1',
+    'ACQ_FIRST_IN_ENCODE_STEP2',
+    'ACQ_LAST_IN_ENCODE_STEP2',
+    'ACQ_FIRST_IN_AVERAGE',
+    'ACQ_LAST_IN_AVERAGE',
+    'ACQ_FIRST_IN_SLICE',
+    'ACQ_LAST_IN_SLICE',
+    'ACQ_FIRST_IN_CONTRAST',
+    'ACQ_LAST_IN_CONTRAST',
+    'ACQ_FIRST_IN_PHASE',
+    'ACQ_LAST_IN_PHASE',
+    'ACQ_FIRST_IN_REPETITION',
+    'ACQ_LAST_IN_REPETITION',
+    'ACQ_FIRST_IN_SET',
+    'ACQ_LAST_IN_SET',
+    'ACQ_FIRST_IN_SEGMENT',
+    'ACQ_LAST_IN_SEGMENT',
+    'ACQ_IS_NOISE_MEASUREMENT',
+    'ACQ_IS_PARALLEL_CALIBRATION',
+    'ACQ_IS_PARALLEL_CALIBRATION_AND_IMAGING',
+    'ACQ_IS_REVERSE',
+    'ACQ_IS_NAVIGATION_DATA',
+    'ACQ_IS_PHASECORR_DATA',
+    'ACQ_LAST_IN_MEASUREMENT',
+    'ACQ_IS_HPFEEDBACK_DATA',
+    'ACQ_IS_DUMMYSCAN_DATA',
+    'ACQ_IS_RTFEEDBACK_DATA',
+    'ACQ_IS_SURFACECOILCORRECTIONSCAN_DATA',
+    'ACQ_USER1',
+    'ACQ_USER2',
+    'ACQ_USER3',
+    'ACQ_USER4',
+    'ACQ_USER5',
+    'ACQ_USER6',
+    'ACQ_USER7',
+    'ACQ_USER8',
+)
